+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Notable changes to this project.
 
 ## [dev]
 - added `friends` report - showing correlation of a user's live auroc to other users
+- fix `payments` report for rounds >= 158
 
 ## [0.2.2] - 2019-06-26
 - create cache directory in current working directory